Enable git-lfs tracking and update sprite pixels per unit
Git LFS will track *.uasset and *.umap files so that our repository
doesn't bloat.

Set the PPU to 3.2 so that a 32x32 pixel sprite aligns with the unreal
engine grid / snapping functionality nicely:

    * A 32 pixel sprite fits into a 10x10 grid of world units.  We may
    tweak this later if rendering to a target resolution comes with
    interpolation issues, or if e.g. we prefer 0.32 for finer-grained
    placement control (32 pixels in 100x100 grid).

Note: We've dropped an animation into the level to preview the sprite
size on the world grid.

Note: Normally I would split this commit into one for git-lfs, and one
for PPU, but hey, the history doesn't have to be that clean just for me.