Implement preliminary hero movement
The pawn can now walk up, down, left, and right.  Flipbook animations
reflect this.

Animations should swap between idle and walking states appropriately,
although this was not yet explicitly tested.

Note: There is currently a bug in that if both axes are pressed forward
(say, up and right are both at 1.0), events will fire each frame to
toggle the flipbook back and forth from north to east to north to east,
etc.  This will cause the flipbook for the current walking direction
to visually lock on frame 0, because it is constantly restarted when the
flipbook toggles.

Note: EDirection was renamed to EBearing for north-south-east-west
directions.