Fix bug in player walking logic
Now the selected flipbook gets updated at most once per frame.  So we
won't have the flipbook switching back and forth and resetting to
time 0 when two input axes are pressed, but the final walking direction
does not change within a frame.

There are still issues in terms of one axis (e.g. north/south vs.
east/west) being preferred over the other when two axes are pressed, but
that is a separate fix for later.